# openapi/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Minimarket Plus API",
    "description": "API REST del sistema backend Minimarket Plus. Gestiona productos, categorías, carritos, inventario, ventas y usuarios con roles (cliente, cajero, administrador). Todos los endpoints de negocio requieren autenticación HTTP Basic; use el botón Authorize (credenciales de demostración en el README).",
    "contact": {
      "name": "Equipo Minimarket Plus - PBY2202",
      "email": "karla.santigut@gmail.com"
    },
    "license": {
      "name": "Uso académico - Duoc UC"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Entorno local de desarrollo"
    }
  ],
  "security": [
    {
      "basicAuth": []
    }
  ],
...
  "paths": {
...
    "/api/carritos/{id}": {
      "get": {
        "tags": [
          "Carritos"
        ],
        "summary": "Obtener un ítem de carrito por ID",
        "description": "Busca un ítem de carrito específico por su identificador único.",
        "operationId": "obtenerCarritoPorId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del ítem de carrito",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un ítem con el ID indicado"
          },
          "200": {
            "description": "Ítem encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Carrito"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Carritos"
        ],
        "summary": "Actualizar un ítem del carrito",
        "description": "Modifica el producto o la cantidad de un ítem de carrito existente.",
        "operationId": "actualizarCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del ítem a actualizar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "requestBody": {
          "description": "Nuevos datos del ítem de carrito",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carrito"
              },
              "examples": {
                "itemActualizado": {
                  "description": "itemActualizado",
                  "value": {
                    "usuario": {
                      "id": 3
                    },
                    "producto": {
                      "id": 1
                    },
                    "cantidad": 3
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Ítem actualizado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Carrito"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un ítem con el ID indicado"
          }
        }
      },
      "delete": {
        "tags": [
          "Carritos"
        ],
        "summary": "Eliminar un producto del carrito",
        "description": "Elimina el ítem de carrito identificado por el ID.",
        "operationId": "eliminarProductoDelCarrito",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identificador único del ítem a eliminar",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 1
          }
        ],
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "404": {
            "description": "No existe un ítem con el ID indicado"
          },
          "204": {
            "description": "Ítem eliminado correctamente"
          }
        }
      }
    },
...
    "/api/carritos": {
      "get": {
        "tags": [
          "Carritos"
        ],
        "summary": "Listar todos los ítems de carrito",
        "description": "Retorna todos los ítems de carrito registrados, con su usuario, producto y cantidad.",
        "operationId": "listarCarrito",
        "responses": {
          "401": {
            "description": "No autenticado"
          },
          "200": {
            "description": "Listado obtenido correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Carrito"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Carritos"
        ],
        "summary": "Agregar un producto al carrito",
        "description": "Crea un ítem de carrito asociando un usuario, un producto existente y la cantidad deseada.",
        "operationId": "agregarProductoAlCarrito",
        "requestBody": {
          "description": "Ítem de carrito a crear (usuario y producto deben existir)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Carrito"
              },
              "examples": {
                "nuevoItemCarrito": {
                  "description": "nuevoItemCarrito",
                  "value": {
                    "usuario": {
                      "id": 3
                    },
                    "producto": {
                      "id": 1
                    },
                    "cantidad": 3
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Producto agregado al carrito correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Carrito"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Rol": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del rol",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre del rol",
            "example": "ROLE_ADMIN"
          }
        },
        "description": "Rol de autorización dentro del sistema"
      },
      "Usuario": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del usuario",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario único para autenticación",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario (solo escritura, no se retorna en respuestas)",
            "writeOnly": true,
            "example": "admin123"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "description": "Roles asignados al usuario",
            "items": {
              "$ref": "#/components/schemas/Rol"
            }
          }
        },
        "description": "Usuario del sistema con sus roles asociados (cliente, cajero, administrador)"
      },
      "Categoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único de la categoría",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre único de la categoría",
            "example": "Bebidas"
          }
        },
        "description": "Categoría para clasificar los productos del minimarket"
      },
      "Producto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del producto",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre comercial del producto",
            "example": "Coca-Cola 1.5L"
          },
          "precio": {
            "type": "number",
            "description": "Precio unitario en pesos chilenos",
            "format": "double",
            "example": 1890
          },
          "stock": {
            "type": "integer",
            "description": "Unidades disponibles en bodega",
            "format": "int32",
            "example": 24
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria"
          }
        },
        "description": "Producto del catálogo del minimarket"
      },
...
      "Carrito": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Identificador único del ítem del carrito",
            "format": "int64",
            "readOnly": true,
            "example": 1
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "description": "Cantidad de unidades del producto",
            "format": "int32",
            "example": 3
          }
        },
        "description": "Ítem del carrito de compras: asocia un producto y su cantidad a un usuario"
      }
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "description": "Autenticación HTTP Basic contra los usuarios registrados en H2",
        "name": "basicAuth",
        "scheme": "basic"
      }
    }
  }
}